// eonev/src/models/Event.ts
export default class Event {
  id: string | undefined;
  title: string | undefined;
  description: string | undefined;
  link: string | undefined;
  closed: Date | undefined;
  categoryId: Number | undefined;
  lastGeometryDate: Date | undefined;
  lastGeometryX: Number | undefined;
  lastGeometryY: Number | undefined;
}